+++ src/main/org/firebirdsql/management/TraceManager.java
@@ -39,7 +39,6 @@
 
 package org.firebirdsql.management;
 
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.sql.SQLException;
 
@@ -99,12 +98,16 @@ public interface TraceManager extends ServiceManager {
     void listTraceSessions() throws SQLException;
     
 	/**
-	 * Loads a configuration from the specified fileName
+	 * Loads a configuration from the specified fileName using the default character set
 	 *
-	 * @throws FileNotFoundException
+     * @param fileName
+     *         File name
 	 * @throws IOException
+     * @deprecated We suggest you use standard Java functionality instead (eg
+     * {@code new String(Files.readAllBytes(Paths.get(fileName)), Charset.defaultCharset())}.
+     * This method will be removed in Jaybird 4.
 	 */
-    String loadConfigurationFromFile(String fileName) throws FileNotFoundException, IOException;
+    String loadConfigurationFromFile(String fileName) throws IOException;
     
     /**
      * Gets the sessionId for the given name.